refactor: move replicas splitting into a separate function
